Makes first steps of an EMAP cracker + Bug fixes
made a prototype EMAP cracker, then made the stage1and2 function to perform the first parts of finding the secret values. Also made changes to the Fp function, to make it easier to debug. Also made changes to EMAP to make it more consistent.

WARNING: currently stage1and2 gets the wrong values for n1 and n2 somehow. Probably related to the truth table for determining all the bits of n2, but unsure for n1's problem. Could be an issue with the EMAP functions.